Add CQRS pattern
