+++ lib/constants.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+# Absolute repo root used by SQLite adapter path construction.
+REPO_ROOT = str(Path(__file__).resolve().parent.parent)
