+++ simulation/core/__init__.py
@@ -12,6 +12,14 @@ def __getattr__(name: str):
         from .dependencies import create_engine
 
         return create_engine
+    if name == "SimulationQueryService":
+        from .query_service import SimulationQueryService
+
+        return SimulationQueryService
+    if name == "SimulationCommandService":
+        from .command_service import SimulationCommandService
+
+        return SimulationCommandService
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
@@ -21,4 +29,6 @@ __all__ = [
     "SimulationError",
     "InsufficientAgentsError",
     "create_engine",  # type: ignore[attr-defined]  # Lazy-loaded via __getattr__
+    "SimulationQueryService",  # type: ignore[attr-defined]  # Lazy-loaded via __getattr__
+    "SimulationCommandService",  # type: ignore[attr-defined]  # Lazy-loaded via __getattr__
 ]
