+++ tests/simulation/core/test_engine.py
@@ -1,26 +1,23 @@
-"""Tests for simulation.core.engine module."""
+"""Facade tests for simulation.core.engine module."""
 
-from unittest.mock import Mock, patch
+from unittest.mock import Mock
 
 import pytest
 
-from db.exceptions import RunNotFoundError, RunStatusUpdateError
 from db.repositories.feed_post_repository import FeedPostRepository
 from db.repositories.generated_bio_repository import GeneratedBioRepository
 from db.repositories.generated_feed_repository import GeneratedFeedRepository
 from db.repositories.profile_repository import ProfileRepository
 from db.repositories.run_repository import RunRepository
+from simulation.core.command_service import SimulationCommandService
 from simulation.core.engine import SimulationEngine
 from simulation.core.models.agents import SocialMediaAgent
-from simulation.core.models.feeds import GeneratedFeed
-from simulation.core.models.posts import BlueskyFeedPost
 from simulation.core.models.runs import Run, RunStatus
-from simulation.core.models.turns import TurnData, TurnResult
+from simulation.core.query_service import SimulationQueryService
 
 
 @pytest.fixture
-def mock_repos():
-    """Fixture that creates and returns a dictionary of mock repositories."""
+def deps():
     return {
         "run_repo": Mock(spec=RunRepository),
         "profile_repo": Mock(spec=ProfileRepository),
@@ -31,851 +28,87 @@ def mock_repos():
 
 
 @pytest.fixture
-def mock_agent_factory():
-    """Fixture that provides a mock agent factory."""
+def agent_factory():
     factory = Mock()
-    factory.side_effect = lambda num_agents: [
-        SocialMediaAgent(f"agent{i}.bsky.social") for i in range(num_agents)
-    ]
+    factory.return_value = [SocialMediaAgent("agent1.bsky.social")]
     return factory
 
 
 @pytest.fixture
-def engine(mock_repos, mock_agent_factory):
-    """Fixture that creates and returns a SimulationEngine with mock repositories."""
+def query_service():
+    return Mock(spec=SimulationQueryService)
+
+
+@pytest.fixture
+def command_service():
+    return Mock(spec=SimulationCommandService)
+
+
+@pytest.fixture
+def engine(deps, agent_factory, query_service, command_service):
     return SimulationEngine(
-        run_repo=mock_repos["run_repo"],
-        profile_repo=mock_repos["profile_repo"],
-        feed_post_repo=mock_repos["feed_post_repo"],
-        generated_bio_repo=mock_repos["generated_bio_repo"],
-        generated_feed_repo=mock_repos["generated_feed_repo"],
-        agent_factory=mock_agent_factory,
+        run_repo=deps["run_repo"],
+        profile_repo=deps["profile_repo"],
+        feed_post_repo=deps["feed_post_repo"],
+        generated_bio_repo=deps["generated_bio_repo"],
+        generated_feed_repo=deps["generated_feed_repo"],
+        agent_factory=agent_factory,
+        query_service=query_service,
+        command_service=command_service,
     )
 
 
-@pytest.fixture
-def sample_run():
-    """Fixture providing a standard Run object for tests."""
-    return Run(
-        run_id="run_123",
-        created_at="2024_01_01-12:00:00",
-        total_turns=10,
-        total_agents=5,
-        started_at="2024_01_01-12:00:00",
-        status=RunStatus.RUNNING,
-        completed_at=None,
-    )
+class TestSimulationEngineCompatibility:
+    def test_keeps_dependency_attributes(self, engine, deps, agent_factory):
+        assert engine.run_repo is deps["run_repo"]
+        assert engine.profile_repo is deps["profile_repo"]
+        assert engine.feed_post_repo is deps["feed_post_repo"]
+        assert engine.generated_bio_repo is deps["generated_bio_repo"]
+        assert engine.generated_feed_repo is deps["generated_feed_repo"]
+        assert engine.agent_factory is agent_factory
 
 
-@pytest.fixture
-def sample_feed1(sample_run):
-    """Fixture providing a standard GeneratedFeed for agent1."""
-    return GeneratedFeed(
-        feed_id="feed_1",
-        run_id=sample_run.run_id,
-        turn_number=0,
-        agent_handle="agent1.bsky.social",
-        post_uris=["uri1", "uri2"],
-        created_at="2024_01_01-12:00:00",
-    )
+class TestSimulationEngineDelegation:
+    def test_delegates_query_methods(self, engine, query_service):
+        engine.get_run("run_123")
+        engine.list_runs()
+        engine.get_turn_metadata("run_123", 0)
+        engine.get_turn_data("run_123", 0)
 
+        query_service.get_run.assert_called_once_with("run_123")
+        query_service.list_runs.assert_called_once()
+        query_service.get_turn_metadata.assert_called_once_with("run_123", 0)
+        query_service.get_turn_data.assert_called_once_with("run_123", 0)
 
-@pytest.fixture
-def sample_feed2(sample_run):
-    """Fixture providing a standard GeneratedFeed for agent2."""
-    return GeneratedFeed(
-        feed_id="feed_2",
-        run_id=sample_run.run_id,
-        turn_number=0,
-        agent_handle="agent2.bsky.social",
-        post_uris=["uri3"],
-        created_at="2024_01_01-12:00:01",
-    )
-
-
-@pytest.fixture
-def sample_posts():
-    """Fixture providing standard BlueskyFeedPost objects for tests."""
-    return [
-        BlueskyFeedPost(
-            id="uri1",
-            uri="uri1",
-            author_display_name="Author 1",
-            author_handle="author1.bsky.social",
-            text="Post 1 text",
-            bookmark_count=0,
-            like_count=5,
-            quote_count=0,
-            reply_count=2,
-            repost_count=1,
+    def test_delegates_command_methods(self, engine, command_service):
+        run = Run(
+            run_id="run_123",
             created_at="2024_01_01-12:00:00",
-        ),
-        BlueskyFeedPost(
-            id="uri2",
-            uri="uri2",
-            author_display_name="Author 2",
-            author_handle="author2.bsky.social",
-            text="Post 2 text",
-            bookmark_count=1,
-            like_count=10,
-            quote_count=0,
-            reply_count=3,
-            repost_count=2,
-            created_at="2024_01_01-12:01:00",
-        ),
-        BlueskyFeedPost(
-            id="uri3",
-            uri="uri3",
-            author_display_name="Author 3",
-            author_handle="author3.bsky.social",
-            text="Post 3 text",
-            bookmark_count=0,
-            like_count=0,
-            quote_count=0,
-            reply_count=0,
-            repost_count=0,
-            created_at="2024_01_01-12:02:00",
-        ),
-    ]
-
-
-@pytest.fixture
-def default_test_params():
-    """Fixture providing default test parameters."""
-    return {
-        "run_id": "run_123",
-        "turn_number": 0,
-    }
-
-
-class TestSimulationEngineGetRun:
-    """Tests for SimulationEngine.get_run method."""
-
-    def test_returns_run_when_found(self, engine, mock_repos):
-        """Test that get_run returns a Run when it exists."""
-        # Arrange
-        run_id = "run_123"
-        expected_run = Run(
-            run_id=run_id,
-            created_at="2024_01_01-12:00:00",
-            total_turns=10,
-            total_agents=5,
+            total_turns=1,
+            total_agents=1,
             started_at="2024_01_01-12:00:00",
             status=RunStatus.RUNNING,
             completed_at=None,
         )
-        mock_repos["run_repo"].get_run.return_value = expected_run
-
-        # Act
-        result = engine.get_run(run_id)
-
-        # Assert
-        assert result is not None
-        assert result == expected_run
-        mock_repos["run_repo"].get_run.assert_called_once_with(run_id)
-
-    def test_returns_none_when_run_not_found(self, engine, mock_repos):
-        """Test that get_run returns None when run does not exist."""
-        # Arrange
-        run_id = "nonexistent_run"
-        mock_repos["run_repo"].get_run.return_value = None
-
-        # Act
-        result = engine.get_run(run_id)
-
-        # Assert
-        assert result is None
-        mock_repos["run_repo"].get_run.assert_called_once_with(run_id)
-
-    def test_raises_value_error_for_empty_run_id(self, engine, mock_repos):
-        """Test that get_run raises ValueError for empty run_id."""
-        # Arrange & Act & Assert
-        with pytest.raises(ValueError, match="run_id cannot be empty"):
-            engine.get_run("")
-
-        # Verify repository was not called
-        mock_repos["run_repo"].get_run.assert_not_called()
-
-    def test_raises_value_error_for_whitespace_only_run_id(self, engine, mock_repos):
-        """Test that get_run raises ValueError for whitespace-only run_id."""
-        # Arrange & Act & Assert
-        with pytest.raises(ValueError, match="run_id cannot be empty"):
-            engine.get_run("   ")
-
-        # Verify repository was not called
-        mock_repos["run_repo"].get_run.assert_not_called()
-
-    def test_raises_value_error_for_none_run_id(self, engine, mock_repos):
-        """Test that get_run raises ValueError for None run_id."""
-        # Arrange & Act & Assert
-        with pytest.raises(ValueError, match="run_id cannot be empty"):
-            engine.get_run(None)  # type: ignore
-
-        # Verify repository was not called
-        mock_repos["run_repo"].get_run.assert_not_called()
-
-    def test_repository_exceptions_propagate(self, engine, mock_repos):
-        """Test that repository exceptions propagate without wrapping."""
-        # Arrange
-        run_id = "run_123"
-        original_error = RunNotFoundError(run_id)
-        mock_repos["run_repo"].get_run.side_effect = original_error
-
-        # Act & Assert
-        with pytest.raises(RunNotFoundError) as exc_info:
-            engine.get_run(run_id)
-
-        # Verify it's the same exception (not wrapped)
-        assert exc_info.value is original_error
-        mock_repos["run_repo"].get_run.assert_called_once_with(run_id)
-
-
-class TestSimulationEngineListRuns:
-    """Tests for SimulationEngine.list_runs method."""
-
-    def test_returns_list_of_runs(self, engine, mock_repos):
-        """Test that list_runs returns a list of runs."""
-        # Arrange
-        expected_runs = [
-            Run(
-                run_id="run_1",
-                created_at="2024_01_01-12:00:00",
-                total_turns=10,
-                total_agents=5,
-                started_at="2024_01_01-12:00:00",
-                status=RunStatus.RUNNING,
-                completed_at=None,
-            ),
-            Run(
-                run_id="run_2",
-                created_at="2024_01_02-12:00:00",
-                total_turns=20,
-                total_agents=10,
-                started_at="2024_01_02-12:00:00",
-                status=RunStatus.COMPLETED,
-                completed_at="2024_01_02-13:00:00",
-            ),
-        ]
-        mock_repos["run_repo"].list_runs.return_value = expected_runs
-
-        # Act
-        result = engine.list_runs()
-
-        # Assert
-        assert result == expected_runs
-        assert len(result) == 2
-        mock_repos["run_repo"].list_runs.assert_called_once()
-
-    def test_returns_empty_list_when_no_runs_exist(self, engine, mock_repos):
-        """Test that list_runs returns empty list when no runs exist."""
-        # Arrange
-        mock_repos["run_repo"].list_runs.return_value = []
-
-        # Act
-        result = engine.list_runs()
-
-        # Assert
-        assert result == []
-        assert len(result) == 0
-        mock_repos["run_repo"].list_runs.assert_called_once()
-
-    def test_repository_exceptions_propagate(self, engine, mock_repos):
-        """Test that repository exceptions propagate without wrapping."""
-        # Arrange
-        original_error = RuntimeError("Database connection failed")
-        mock_repos["run_repo"].list_runs.side_effect = original_error
-
-        # Act & Assert
-        with pytest.raises(RuntimeError) as exc_info:
-            engine.list_runs()
-
-        # Verify it's the same exception (not wrapped)
-        assert exc_info.value is original_error
-        mock_repos["run_repo"].list_runs.assert_called_once()
-
-
-class TestSimulationEngineGetTurnData:
-    """Tests for SimulationEngine.get_turn_data method."""
-
-    def test_returns_turn_data_with_feeds_and_posts(
-        self,
-        engine,
-        mock_repos,
-        sample_run,
-        sample_feed1,
-        sample_feed2,
-        sample_posts,
-        default_test_params,
-    ):
-        """Test that get_turn_data returns TurnData with correct feeds and posts."""
-        # Arrange
-        run_id = default_test_params["run_id"]
-        turn_number = default_test_params["turn_number"]
-
-        mock_repos["run_repo"].get_run.return_value = sample_run
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.return_value = [
-            sample_feed1,
-            sample_feed2,
-        ]
-        mock_repos["feed_post_repo"].read_feed_posts_by_uris.return_value = sample_posts
-
-        # Act
-        result = engine.get_turn_data(run_id, turn_number)
-
-        # Assert
-        assert result is not None
-        assert isinstance(result, TurnData)
-        assert result.turn_number == turn_number
-        assert result.agents == []
-        assert result.actions == {}
-        assert len(result.feeds) == 2
-        assert "agent1.bsky.social" in result.feeds
-        assert "agent2.bsky.social" in result.feeds
-        assert len(result.feeds["agent1.bsky.social"]) == 2
-        assert len(result.feeds["agent2.bsky.social"]) == 1
-        assert result.feeds["agent1.bsky.social"][0].uri == "uri1"
-        assert result.feeds["agent1.bsky.social"][1].uri == "uri2"
-        assert result.feeds["agent2.bsky.social"][0].uri == "uri3"
-
-        # Verify repository calls
-        mock_repos["run_repo"].get_run.assert_called_once_with(run_id)
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.assert_called_once_with(
-            run_id, turn_number
-        )
-        mock_repos["feed_post_repo"].read_feed_posts_by_uris.assert_called_once()
-        # Verify batch query was called with all URIs
-        call_args = mock_repos["feed_post_repo"].read_feed_posts_by_uris.call_args[0][0]
-        assert set(call_args) == {"uri1", "uri2", "uri3"}
-
-    def test_returns_none_when_turn_does_not_exist(
-        self, engine, mock_repos, sample_run, default_test_params
-    ):
-        """Test that get_turn_data returns None when turn doesn't exist (no feeds)."""
-        # Arrange
-        run_id = default_test_params["run_id"]
-        turn_number = default_test_params["turn_number"]
-
-        mock_repos["run_repo"].get_run.return_value = sample_run
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.return_value = []
-
-        # Act
-        result = engine.get_turn_data(run_id, turn_number)
-
-        # Assert
-        assert result is None
-        mock_repos["run_repo"].get_run.assert_called_once_with(run_id)
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.assert_called_once_with(
-            run_id, turn_number
-        )
-        # Should not call feed_post_repo when no feeds
-        mock_repos["feed_post_repo"].read_feed_posts_by_uris.assert_not_called()
-
-    def test_raises_value_error_for_empty_run_id(self, engine, mock_repos):
-        """Test that get_turn_data raises ValueError for empty run_id."""
-        # Arrange & Act & Assert
-        with pytest.raises(ValueError, match="run_id cannot be empty"):
-            engine.get_turn_data("", 0)
-
-        # Verify repositories were not called
-        mock_repos["run_repo"].get_run.assert_not_called()
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.assert_not_called()
-
-    def test_raises_value_error_for_whitespace_only_run_id(self, engine, mock_repos):
-        """Test that get_turn_data raises ValueError for whitespace-only run_id."""
-        # Arrange & Act & Assert
-        with pytest.raises(ValueError, match="run_id cannot be empty"):
-            engine.get_turn_data("   ", 0)
-
-        # Verify repositories were not called
-        mock_repos["run_repo"].get_run.assert_not_called()
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.assert_not_called()
-
-    def test_raises_value_error_for_negative_turn_number(self, engine, mock_repos):
-        """Test that get_turn_data raises ValueError for negative turn_number."""
-        # Arrange & Act & Assert
-        with pytest.raises(ValueError, match="turn_number cannot be negative"):
-            engine.get_turn_data("run_123", -1)
-
-        # Verify repositories were not called
-        mock_repos["run_repo"].get_run.assert_not_called()
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.assert_not_called()
-
-    def test_raises_value_error_for_none_run_id(self, engine, mock_repos):
-        """Test that get_turn_data raises ValueError for None run_id."""
-        # Arrange & Act & Assert
-        with pytest.raises(ValueError, match="run_id cannot be empty"):
-            engine.get_turn_data(None, 0)  # type: ignore
-
-        # Verify repositories were not called
-        mock_repos["run_repo"].get_run.assert_not_called()
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.assert_not_called()
-
-    def test_raises_value_error_for_none_turn_number(self, engine, mock_repos):
-        """Test that get_turn_data raises ValueError for None turn_number."""
-        # Arrange & Act & Assert
-        with pytest.raises(ValueError):
-            engine.get_turn_data("run_123", None)  # type: ignore
-
-        # Verify repositories were not called
-        mock_repos["run_repo"].get_run.assert_not_called()
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.assert_not_called()
-
-    def test_raises_run_not_found_error_when_run_does_not_exist(
-        self, engine, mock_repos
-    ):
-        """Test that get_turn_data raises RunNotFoundError for non-existent run."""
-        # Arrange
-        run_id = "nonexistent_run"
-        turn_number = 0
-        mock_repos["run_repo"].get_run.return_value = None
-
-        # Act & Assert
-        with pytest.raises(RunNotFoundError) as exc_info:
-            engine.get_turn_data(run_id, turn_number)
-
-        assert exc_info.value.run_id == run_id
-        mock_repos["run_repo"].get_run.assert_called_once_with(run_id)
-        # Should not call other repositories
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.assert_not_called()
-        mock_repos["feed_post_repo"].read_feed_posts_by_uris.assert_not_called()
-
-    def test_handles_missing_posts_gracefully(
-        self, engine, mock_repos, sample_run, sample_posts, default_test_params
-    ):
-        """Test that get_turn_data handles missing posts gracefully (skips missing, returns partial feeds)."""
-        # Arrange
-        run_id = default_test_params["run_id"]
-        turn_number = default_test_params["turn_number"]
-
-        feed = GeneratedFeed(
-            feed_id="feed_1",
-            run_id=run_id,
-            turn_number=turn_number,
-            agent_handle="agent1.bsky.social",
-            post_uris=[
-                "uri1",
-                "uri2",
-                "missing_uri",
-            ],  # uri2 and missing_uri have no corresponding post objects in the repository (only uri1 does)
-            created_at="2024_01_01-12:00:00",
-        )
-
-        # Only uri1 exists, uri2 and missing_uri don't
-        mock_repos["run_repo"].get_run.return_value = sample_run
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.return_value = [feed]
-        mock_repos["feed_post_repo"].read_feed_posts_by_uris.return_value = [
-            sample_posts[0]
-        ]  # Only first post
-
-        # Act
-        result = engine.get_turn_data(run_id, turn_number)
-
-        # Assert
-        assert result is not None
-        assert len(result.feeds["agent1.bsky.social"]) == 1  # Only uri1 found
-        assert result.feeds["agent1.bsky.social"][0].uri == "uri1"
-        # Verify batch query was called with all URIs (including missing ones)
-        call_args = mock_repos["feed_post_repo"].read_feed_posts_by_uris.call_args[0][0]
-        assert set(call_args) == {"uri1", "uri2", "missing_uri"}
-
-    def test_handles_empty_feeds_when_all_posts_missing(
-        self, engine, mock_repos, sample_run, default_test_params
-    ):
-        """Test that get_turn_data returns TurnData with empty feeds dict when all posts missing."""
-        # Arrange
-        run_id = default_test_params["run_id"]
-        turn_number = default_test_params["turn_number"]
-
-        feed = GeneratedFeed(
-            feed_id="feed_1",
-            run_id=run_id,
-            turn_number=turn_number,
-            agent_handle="agent1.bsky.social",
-            post_uris=["missing_uri1", "missing_uri2"],
-            created_at="2024_01_01-12:00:00",
-        )
-
-        mock_repos["run_repo"].get_run.return_value = sample_run
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.return_value = [feed]
-        # No posts found
-        mock_repos["feed_post_repo"].read_feed_posts_by_uris.return_value = []
-
-        # Act
-        result = engine.get_turn_data(run_id, turn_number)
-
-        # Assert
-        assert result is not None
-        assert result.turn_number == turn_number
-        assert result.feeds["agent1.bsky.social"] == []  # Empty list, not None
-        # Turn exists but data is incomplete
-
-    def test_handles_multiple_feeds_with_overlapping_uris(
-        self, engine, mock_repos, sample_run, default_test_params
-    ):
-        """Test that get_turn_data handles multiple feeds with overlapping post URIs correctly."""
-        # Arrange
-        run_id = default_test_params["run_id"]
-        turn_number = default_test_params["turn_number"]
-
-        # Two feeds with overlapping URIs
-        feed1 = GeneratedFeed(
-            feed_id="feed_1",
-            run_id=run_id,
-            turn_number=turn_number,
-            agent_handle="agent1.bsky.social",
-            post_uris=["uri1", "uri2"],  # uri2 is shared
-            created_at="2024_01_01-12:00:00",
-        )
-        feed2 = GeneratedFeed(
-            feed_id="feed_2",
-            run_id=run_id,
-            turn_number=turn_number,
-            agent_handle="agent2.bsky.social",
-            post_uris=["uri2", "uri3"],  # uri2 is shared
-            created_at="2024_01_01-12:01:00",
-        )
-
-        post1 = BlueskyFeedPost(
-            id="uri1",
-            uri="uri1",
-            author_display_name="Author 1",
-            author_handle="author1.bsky.social",
-            text="Post 1",
-            bookmark_count=0,
-            like_count=0,
-            quote_count=0,
-            reply_count=0,
-            repost_count=0,
-            created_at="2024_01_01-12:00:00",
-        )
-        post2 = BlueskyFeedPost(
-            id="uri2",
-            uri="uri2",
-            author_display_name="Author 2",
-            author_handle="author2.bsky.social",
-            text="Post 2",
-            bookmark_count=0,
-            like_count=0,
-            quote_count=0,
-            reply_count=0,
-            repost_count=0,
-            created_at="2024_01_01-12:01:00",
-        )
-        post3 = BlueskyFeedPost(
-            id="uri3",
-            uri="uri3",
-            author_display_name="Author 3",
-            author_handle="author3.bsky.social",
-            text="Post 3",
-            bookmark_count=0,
-            like_count=0,
-            quote_count=0,
-            reply_count=0,
-            repost_count=0,
-            created_at="2024_01_01-12:02:00",
-        )
-
-        mock_repos["run_repo"].get_run.return_value = sample_run
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.return_value = [
-            feed1,
-            feed2,
-        ]
-        mock_repos["feed_post_repo"].read_feed_posts_by_uris.return_value = [
-            post1,
-            post2,
-            post3,
-        ]
-
-        # Act
-        result = engine.get_turn_data(run_id, turn_number)
-
-        # Assert
-        assert result is not None
-        assert len(result.feeds) == 2
-        assert len(result.feeds["agent1.bsky.social"]) == 2  # uri1, uri2
-        assert len(result.feeds["agent2.bsky.social"]) == 2  # uri2, uri3
-        # Verify batch query was called with unique URIs only (uri1, uri2, uri3)
-        call_args = mock_repos["feed_post_repo"].read_feed_posts_by_uris.call_args[0][0]
-        assert set(call_args) == {"uri1", "uri2", "uri3"}
-        assert len(call_args) == 3  # No duplicates
-
-    def test_repository_exceptions_propagate(
-        self, engine, mock_repos, sample_run, default_test_params
-    ):
-        """Test that repository exceptions propagate without wrapping."""
-        # Arrange
-        run_id = default_test_params["run_id"]
-        turn_number = default_test_params["turn_number"]
-
-        original_error = RuntimeError("Database connection failed")
-        mock_repos["run_repo"].get_run.return_value = sample_run
-        mock_repos[
-            "generated_feed_repo"
-        ].read_feeds_for_turn.side_effect = original_error
-
-        # Act & Assert
-        with pytest.raises(RuntimeError) as exc_info:
-            engine.get_turn_data(run_id, turn_number)
-
-        # Verify it's the same exception (not wrapped)
-        assert exc_info.value is original_error
-        mock_repos["run_repo"].get_run.assert_called_once_with(run_id)
-        mock_repos["generated_feed_repo"].read_feeds_for_turn.assert_called_once_with(
-            run_id, turn_number
-        )
-
-
-class TestSimulationEngineUpdateRunStatus:
-    """Tests for SimulationEngine.update_run_status method."""
-
-    def test_updates_status_successfully(self, engine, mock_repos, sample_run):
-        """Test that update_run_status writes the requested status."""
-        # Arrange
-        status = RunStatus.COMPLETED
-        mock_repos["run_repo"].update_run_status.return_value = None
-
-        # Act
-        engine.update_run_status(sample_run, status)
-
-        # Assert
-        mock_repos["run_repo"].update_run_status.assert_called_once_with(
-            sample_run.run_id, status
-        )
-
-    def test_retries_then_succeeds(self, engine, mock_repos, sample_run):
-        """Test that update_run_status retries transient failures."""
-        mock_repos["run_repo"].update_run_status.side_effect = [
-            RunStatusUpdateError(sample_run.run_id, "first"),
-            RunStatusUpdateError(sample_run.run_id, "second"),
-            None,
-        ]
-
-        with patch("simulation.core.engine.time.sleep") as mock_sleep:
-            engine.update_run_status(sample_run, RunStatus.RUNNING)
-
-        calls = mock_repos["run_repo"].update_run_status.call_args_list
-        assert len(calls) == 3
-        assert calls[0][0] == (sample_run.run_id, RunStatus.RUNNING)
-        assert calls[1][0] == (sample_run.run_id, RunStatus.RUNNING)
-        assert calls[2][0] == (sample_run.run_id, RunStatus.RUNNING)
-        assert mock_sleep.call_count == 2
-
-    def test_retry_exhausted_marks_failed_and_raises(
-        self, engine, mock_repos, sample_run
-    ):
-        """If status update fails 3 times, mark FAILED best-effort then raise."""
-        mock_repos["run_repo"].update_run_status.side_effect = [
-            RunStatusUpdateError(sample_run.run_id, "first"),
-            RunStatusUpdateError(sample_run.run_id, "second"),
-            RunStatusUpdateError(sample_run.run_id, "third"),
-            None,  # FAILED best-effort
-        ]
-
-        with patch("simulation.core.engine.time.sleep") as mock_sleep:
-            with pytest.raises(RunStatusUpdateError):
-                engine.update_run_status(sample_run, RunStatus.COMPLETED)
-
-        calls = mock_repos["run_repo"].update_run_status.call_args_list
-        assert len(calls) == 4
-        assert calls[0][0] == (sample_run.run_id, RunStatus.COMPLETED)
-        assert calls[1][0] == (sample_run.run_id, RunStatus.COMPLETED)
-        assert calls[2][0] == (sample_run.run_id, RunStatus.COMPLETED)
-        assert calls[3][0] == (sample_run.run_id, RunStatus.FAILED)
-        assert mock_sleep.call_count == 2
-
-    def test_retry_exhausted_for_failed_does_not_retry_failed_fallback(
-        self, engine, mock_repos, sample_run
-    ):
-        """If FAILED update itself exhausts retries, don't do extra fallback call."""
-        mock_repos["run_repo"].update_run_status.side_effect = [
-            RunStatusUpdateError(sample_run.run_id, "first"),
-            RunStatusUpdateError(sample_run.run_id, "second"),
-            RunStatusUpdateError(sample_run.run_id, "third"),
-        ]
-
-        with patch("simulation.core.engine.time.sleep"):
-            with pytest.raises(RunStatusUpdateError):
-                engine.update_run_status(sample_run, RunStatus.FAILED)
-
-        calls = mock_repos["run_repo"].update_run_status.call_args_list
-        assert len(calls) == 3
-        assert all(call[0] == (sample_run.run_id, RunStatus.FAILED) for call in calls)
-
-
-class TestSimulationEngineExecuteRun:
-    """Tests for SimulationEngine.execute_run method."""
-
-    def _make_run(self, run_id: str = "run_exec_1", total_turns: int = 2) -> Run:
-        return Run(
-            run_id=run_id,
-            created_at="2024_01_01-12:00:00",
-            total_turns=total_turns,
-            total_agents=5,
-            started_at="2024_01_01-12:00:00",
-            status=RunStatus.RUNNING,
-            completed_at=None,
-        )
-
-    def test_success_with_running_retry_and_completed(
-        self, engine, mock_repos, mock_agent_factory
-    ):
-        """execute_run should retry RUNNING update, run all turns, then set COMPLETED."""
-        run = self._make_run(total_turns=2)
-        mock_repos["run_repo"].create_run.return_value = run
-
-        # First two RUNNING updates fail, third succeeds; final COMPLETED succeeds
-        mock_repos["run_repo"].update_run_status.side_effect = [
-            RunStatusUpdateError(run.run_id, "first"),
-            RunStatusUpdateError(run.run_id, "second"),
-            None,  # RUNNING success on 3rd try
-            None,  # COMPLETED
-        ]
-
-        # Configure agent factory to return agents
-        mock_agent_factory.return_value = [
-            SocialMediaAgent("agent1.bsky.social"),
-            SocialMediaAgent("agent2.bsky.social"),
-        ]
-
-        with (
-            patch(
-                "simulation.core.engine.SimulationEngine._simulate_turn"
-            ) as mock_sim_turn,
-            patch("simulation.core.engine.time.sleep") as mock_sleep,
-        ):
-            mock_sim_turn.side_effect = [
-                TurnResult(turn_number=0, total_actions={}, execution_time_ms=10),
-                TurnResult(turn_number=1, total_actions={}, execution_time_ms=12),
-            ]
-            mock_sleep.return_value = None  # Avoid delays
-
-            result = engine.execute_run(
-                run_config=type(
-                    "Cfg",
-                    (),
-                    {
-                        "feed_algorithm": "chronological",
-                        "num_agents": 2,
-                        "num_turns": 2,
-                    },
-                )()
-            )
-
-            assert result is run
-            assert mock_sim_turn.call_count == 2
-            # Verify agent_factory was called with correct number of agents
-            mock_agent_factory.assert_called_once_with(2)
-
-            # Verify status updates: 3 attempts to RUNNING, then COMPLETED
-            calls = mock_repos["run_repo"].update_run_status.call_args_list
-            assert len(calls) == 4
-            assert calls[0][0] == (run.run_id, RunStatus.RUNNING)
-            assert calls[1][0] == (run.run_id, RunStatus.RUNNING)
-            assert calls[2][0] == (run.run_id, RunStatus.RUNNING)
-            assert calls[3][0] == (run.run_id, RunStatus.COMPLETED)
-
-    def test_running_retry_exhausted_marks_failed_and_raises(self, engine, mock_repos):
-        """If RUNNING update fails 3 times, mark FAILED best-effort and raise."""
-        run = self._make_run()
-        mock_repos["run_repo"].create_run.return_value = run
-        mock_repos["run_repo"].update_run_status.side_effect = [
-            RunStatusUpdateError(run.run_id, "first"),
-            RunStatusUpdateError(run.run_id, "second"),
-            RunStatusUpdateError(run.run_id, "third"),
-            None,  # FAILED best-effort
-        ]
-
-        with patch("simulation.core.engine.time.sleep") as mock_sleep:
-            mock_sleep.return_value = None
-            with pytest.raises(RunStatusUpdateError):
-                engine.execute_run(
-                    run_config=type(
-                        "Cfg",
-                        (),
-                        {
-                            "feed_algorithm": "chronological",
-                            "num_agents": 2,
-                            "num_turns": 1,
-                        },
-                    )()
-                )
-            calls = mock_repos["run_repo"].update_run_status.call_args_list
-            assert len(calls) == 4
-            assert calls[0][0] == (run.run_id, RunStatus.RUNNING)
-            assert calls[1][0] == (run.run_id, RunStatus.RUNNING)
-            assert calls[2][0] == (run.run_id, RunStatus.RUNNING)
-            assert calls[3][0] == (run.run_id, RunStatus.FAILED)
-
-    def test_agent_creation_failure_marks_failed_and_raises(
-        self, engine, mock_repos, mock_agent_factory
-    ):
-        """If agent creation fails, mark FAILED best-effort and raise."""
-        run = self._make_run()
-        mock_repos["run_repo"].create_run.return_value = run
-        mock_repos["run_repo"].update_run_status.return_value = None  # RUNNING ok
-
-        # Configure agent factory to raise error
-        mock_agent_factory.side_effect = RuntimeError("agent failure")
-
-        with pytest.raises(RuntimeError):
-            engine.execute_run(
-                run_config=type(
-                    "Cfg",
-                    (),
-                    {
-                        "feed_algorithm": "chronological",
-                        "num_agents": 2,
-                        "num_turns": 1,
-                    },
-                )()
-            )
-        calls = mock_repos["run_repo"].update_run_status.call_args_list
-        assert len(calls) == 2
-        assert calls[0][0] == (run.run_id, RunStatus.RUNNING)
-        assert calls[1][0] == (run.run_id, RunStatus.FAILED)
-        # Verify agent_factory was called
-        mock_agent_factory.assert_called_once_with(2)
-
-    def test_turn_failure_marks_failed_and_wraps(
-        self, engine, mock_repos, mock_agent_factory
-    ):
-        """If a turn fails, mark FAILED best-effort and raise wrapped RuntimeError."""
-        run = self._make_run(total_turns=2)
-        mock_repos["run_repo"].create_run.return_value = run
-        mock_repos[
-            "run_repo"
-        ].update_run_status.return_value = (
-            None  # RUNNING ok (and later COMPLETED not reached)
-        )
-
-        # Configure agent factory to return agents
-        mock_agent_factory.return_value = [
-            SocialMediaAgent("agent1.bsky.social"),
-            SocialMediaAgent("agent2.bsky.social"),
-        ]
-
-        with patch("simulation.core.engine.SimulationEngine._simulate_turn") as mock_sim_turn:
-            mock_sim_turn.side_effect = RuntimeError("turn exploded")
-
-            with pytest.raises(RuntimeError) as exc:
-                engine.execute_run(
-                    run_config=type(
-                        "Cfg",
-                        (),
-                        {
-                            "feed_algorithm": "chronological",
-                            "num_agents": 2,
-                            "num_turns": 2,
-                        },
-                    )()
-                )
-            # Wrapped message contains run id and turn number context
-            assert str(exc.value).startswith(
-                f"Failed to complete turn 0 for run {run.run_id}: "
-            )
-            calls = mock_repos["run_repo"].update_run_status.call_args_list
-            assert len(calls) == 2
-            assert calls[0][0] == (run.run_id, RunStatus.RUNNING)
-            assert calls[1][0] == (run.run_id, RunStatus.FAILED)
-            # Verify agent_factory was called
-            mock_agent_factory.assert_called_once_with(2)
+        config = type(
+            "Cfg",
+            (),
+            {
+                "feed_algorithm": "chronological",
+                "num_agents": 1,
+                "num_turns": 1,
+            },
+        )()
+        agents = [SocialMediaAgent("agent1.bsky.social")]
+
+        engine.execute_run(config)
+        engine.update_run_status(run, RunStatus.FAILED)
+        engine.simulate_turn(run, config, 0, agents)
+        engine.simulate_turns(1, run, config, agents)
+        engine.create_agents_for_run(run, config)
+
+        command_service.execute_run.assert_called_once_with(config)
+        command_service.update_run_status.assert_called_once_with(run, RunStatus.FAILED)
+        command_service.simulate_turn.assert_called_once_with(run, config, 0, agents)
+        command_service.simulate_turns.assert_called_once_with(1, run, config, agents)
+        command_service.create_agents_for_run.assert_called_once_with(run, config)
